refactor: Extract transaction helper from component
